chore: migrate benchmarks to ES modules
Converted benchmark scripts from CommonJS to ES modules. Updated package.json scripts and dependencies accordingly, and added "type": "module" for compatibility.

(This commit message was AI-generated.)
