+++ benchmarks/package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "dependencies": {
     "chalk": "^4.1.2",
     "cli-table3": "^0.6.5",
@@ -14,7 +15,7 @@
   "overrides_comment": "@@ Remove",
   "private": "true",
   "scripts": {
-    "backtest": "node backtest.cjs",
-    "benchmark": "node benchmark.cjs"
+    "backtest": "node backtest.js",
+    "benchmark": "node benchmark.js"
   }
 }
